refactor aoti-driven backends
