--- backends/aoti/README.md
@@ -1,2 +1,0 @@
-## Experimental AOTI backend
-Proceed with caution. This is an experimental backend that is not yet ready for production use.
