+++ backends/aoti/common_shims.cpp
@@ -6,13 +6,23 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#include "tensor_attribute.h"
+#include "common_shims.h"
+#include <executorch/runtime/platform/log.h>
+#include <cstdint>
+#include <cstdio>
+#include <fstream>
 #include <iostream>
-#include "utils.h"
+#include <stdexcept>
 
 namespace executorch {
 namespace backends {
 namespace aoti {
+
+namespace internal {
+// Constants for file operations
+const char* const TENSOR_OUTPUT_FILENAME =
+    "/home/gasoonjia/executorch/aoti_intermediate_output.txt";
+} // namespace internal
 
 // Global storage for tensor metadata
 std::unordered_map<Tensor*, std::vector<int64_t>> tensor_to_sizes;
@@ -20,6 +30,7 @@ std::unordered_map<Tensor*, std::vector<int64_t>> tensor_to_strides;
 
 extern "C" {
 
+// Autograd mode functions
 int32_t aoti_torch_grad_mode_is_enabled() {
   // No autograd ever
   return false;
@@ -31,6 +42,7 @@ void aoti_torch_grad_mode_set_enabled(bool enabled) {
   }
 }
 
+// Tensor attribute operations
 AOTITorchError aoti_torch_get_data_ptr(
     AOTITensorHandle tensor,
     void** ret_data_ptr) {
@@ -69,12 +81,6 @@ AOTITorchError aoti_torch_get_dtype(
     int32_t* ret_dtype) {
   *ret_dtype = static_cast<int32_t>(tensor->scalar_type());
 
-  // ASSERTION: Only float32 tensors are supported
-  AOTITorchError dtype_error = validate_dtype(*ret_dtype);
-  if (dtype_error != Error::Ok) {
-    return dtype_error;
-  }
-
   return Error::Ok;
 }
 
@@ -100,13 +106,6 @@ AOTITorchError aoti_torch_get_storage_size(
   throw std::runtime_error("Cannot get storage size on ETensor");
 }
 
-AOTITorchError aoti_torch_get_device_type(
-    AOTITensorHandle tensor,
-    int32_t* ret_device_type) {
-  // All tensors in aoti-cuda delegate are on CUDA
-  *ret_device_type = aoti_torch_device_type_cuda();
-  return Error::Ok;
-}
 
 AOTITorchError aoti_torch_get_device_index(
     AOTITensorHandle tensor,
@@ -121,6 +120,7 @@ AOTITorchError aoti_torch_get_dim(AOTITensorHandle tensor, int64_t* ret_dim) {
   return Error::Ok;
 }
 
+// Device and layout utility functions
 int32_t aoti_torch_device_type_cpu() {
   // Let's say cpu is 0 for ET as well
   return 0;
@@ -132,58 +132,21 @@ __attribute__((__visibility__("default"))) int32_t aoti_torch_layout_strided() {
   return 0;
 }
 
-__attribute__((__visibility__("default"))) int32_t
-aoti_torch_device_type_cuda() {
-  // Let's say cuda is 1 for ET as well
-  return 1;
-}
-
 // Dtype constants - these return the PyTorch dtype codes
 // Currently only float32 is supported, but using robust enum-based approach
 __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_float32() {
-  return static_cast<int32_t>(SupportedDTypes::FLOAT32);
+  return 6; // PyTorch's float32 dtype code
 }
 
-// Future dtype support (commented out for now):
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_bool() {
-//   return static_cast<int32_t>(SupportedDTypes::BOOL);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_uint8() {
-//   return static_cast<int32_t>(SupportedDTypes::UINT8);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_int8() {
-//   return static_cast<int32_t>(SupportedDTypes::INT8);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_int16() {
-//   return static_cast<int32_t>(SupportedDTypes::INT16);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_int32() {
-//   return static_cast<int32_t>(SupportedDTypes::INT32);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_int64() {
-//   return static_cast<int32_t>(SupportedDTypes::INT64);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_float16() {
-//   return static_cast<int32_t>(SupportedDTypes::FLOAT16);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_float64() {
-//   return static_cast<int32_t>(SupportedDTypes::FLOAT64);
-// }
-// 
-// __attribute__((__visibility__("default"))) int32_t aoti_torch_dtype_bfloat16() {
-//   return static_cast<int32_t>(SupportedDTypes::BFLOAT16);
-// }
-
+// Cleanup functions
 void cleanup_tensor_metadata() {
   tensor_to_sizes.clear();
   tensor_to_strides.clear();
+}
+
+void cleanup_aoti_tensor_output() {
+  // Clean up any tensor output related resources
+  // For now this is a no-op, but can be extended if needed
 }
 
 } // extern "C"
