+++ backends/aoti/cuda/TARGETS
@@ -1,0 +1,3 @@
+load("targets.bzl", "define_common_targets")
+
+define_common_targets()
